+++ src/main/java/net/snowflake/client/core/SFSession.java
@@ -143,6 +143,8 @@ public class SFSession extends SFBaseSession {
    */
   private int retryTimeout = 300;
 
+  private int defaultPlatformDetectionTimeoutMs = 200;
+
   private boolean enableClientStoreTemporaryCredential = true;
   private boolean enableClientRequestMfaToken = true;
 
@@ -794,7 +796,16 @@ public class SFSession extends SFBaseSession {
                 : false)
         .setEnableClientStoreTemporaryCredential(enableClientStoreTemporaryCredential)
         .setEnableClientRequestMfaToken(enableClientRequestMfaToken)
-        .setBrowserResponseTimeout(browserResponseTimeout);
+        .setBrowserResponseTimeout(browserResponseTimeout)
+        .setPlatformDetectionTimeoutMs(
+            connectionPropertiesMap.get(SFSessionProperty.PLATFORM_DETECTION_TIMEOUT_MS) != null
+                ? (int) connectionPropertiesMap.get(SFSessionProperty.PLATFORM_DETECTION_TIMEOUT_MS)
+                : defaultPlatformDetectionTimeoutMs)
+        .setDisablePlatformDetection(
+            connectionPropertiesMap.get(SFSessionProperty.DISABLE_PLATFORM_DETECTION) != null
+                ? getBooleanValue(
+                    connectionPropertiesMap.get(SFSessionProperty.DISABLE_PLATFORM_DETECTION))
+                : false); // Default to false (platform detection enabled)
 
     logger.info(
         "Connecting to {} Snowflake domain",